requirements.txt
requests==2.31.0
ollama==0.1.9
mlflow==2.11.0
python-dotenv  # for .env TVH creds
